test: add a pytest test suite
This test suite makes use of the gwacom library and its GObject
bindings, allowing us to write the actual tests in python.

Real devices can be specified as a YAML file in test/devices.
In the tests we can either create custom device or load the correct
Pen/Finger/Pad component from the real device. Devices are then created
as uinput devices [1], with a Monitor class set up to capture any
events.

The test needs to initialize that device and monitor, then play events
through the device and analyze the output.

[1] This unfortunately won't work in containers, like the github CI...

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>